fix: skip migrating read_state of deleted threads
In cs_comment_service, the read_state of a thread is retained even after the thread is deleted
This error is occurring during migration of read_states for old courses that has data in mongodb and has read_state saved for deleted threads.

This commit ensures that the migration process will skip any read_state associated with threads that no longer exist.
close #143