build: Don't branch or tag this in the named release
We only branch and tag top-level backend services,
frontend apps, and non-default Open edX plugins.
This, on the other hand, is a package that gets installed into
edx-platform. We shouldn't tag it, because its `release/ulmo` branch is not guaranteed to line up with the openedx-forum version installed by the
`release/ulmo` branch of edx-platform.

After merging, we should delete the `release/ulmo` tag.